Initial Commit of LiFX.groovy
Initial commit of LiFX.groovy. Current functionality is limited. You must be running LiFX-HTTP on a local server within the network. Settings for the device type can point to the local server address, or an external IP if port forwarding is set up. You can set a specific bulb group to target with the "tag" string, or control all bulbs by entering "all". Work is in progress to add bulb state sync functionality to determine the current value of your specified bulb group and apply the appropriate current values in the HTTP string.